+++ contact/admin.py
@@ -9,7 +9,7 @@ class ContactRequestAdmin(admin.ModelAdmin):
     list_filter = ('contact_type', 'modified', )
 
     def nicer_contact_type(self, obj):
-        return '%s (%s)'.format(obj.get_contact_type_display(), obj.contact_details)
+        return u'{} ({})'.format(obj.get_contact_type_display(), obj.contact_details)
 
     nicer_contact_type.short_description = _(u'Contact via')