+++ contact/forms.py
@@ -8,21 +8,29 @@ from allink_core.allink_base.forms import AllinkBaseModelForm
 from allink_apps.contact.models import ContactRequest
 
 
-class ContactRequestBaseForm(forms.ModelForm):
+class ContactRequestBaseForm(AllinkBaseModelForm):
 
+    salutation = forms.IntegerField(
+        label=_('Salutation'),
+        required=True
+    )
+    email = forms.CharField(
+        label=_(u'Email'),
+        required=False
+    )
     date = forms.DateField(
+            label=_(u'Datum'),
             widget=forms.TextInput(
                 attrs={
                     'class': 'datepicker',
                     'placeholder': _(u'Please choose a date.')
-                    }
-                ),
-            required=False
-            )
+                }),
+        required=False
+    )
 
     class Meta:
         model = None
-        fields = ('first_name', 'last_name', 'email', 'message', 'contact_type', 'date', 'time')
+        fields = ('salutation', 'first_name', 'last_name', 'company_name', 'contact_type', 'email', 'message', 'phone', 'date', 'time')
 
     def clean_date(self):
         date = self.cleaned_data['date']
@@ -34,19 +42,12 @@ class ContactRequestBaseForm(forms.ModelForm):
             raise forms.ValidationError(_(u'This field is required.'))
         return date
 
-    def clean_email(self):
+    def clean_phone(self):
         contact_type = self.cleaned_data.get('contact_type')
-        email = self.cleaned_data['email']
-        if contact_type == ContactRequest.CONTACT_EMAIL and not email:
+        phone = self.cleaned_data['phone']
+        if contact_type == ContactRequest.CONTACT_PHONE and not phone:
             raise forms.ValidationError(_(u'This field is required.'))
-        return email
-
-    def clean_telephone(self):
-        contact_type = self.cleaned_data.get('contact_type')
-        telephone = self.cleaned_data['telephone']
-        if contact_type == ContactRequest.CONTACT_PHONE and not telephone:
-            raise forms.ValidationError(_(u'This field is required.'))
-        return telephone
+        return phone
 
     def clean_time(self):
         contact_type = self.cleaned_data.get('contact_type')
@@ -55,17 +56,16 @@ class ContactRequestBaseForm(forms.ModelForm):
             raise forms.ValidationError(_(u'This field is required.'))
         return time
 
+    def clean_email(self):
+        contact_type = self.cleaned_data.get('contact_type')
+        email = self.cleaned_data['email']
+        if contact_type == ContactRequest.CONTACT_EMAIL and not email:
+            raise forms.ValidationError(_(u'This field is required.'))
+        return email
+
 
 class ContactRequestForm(ContactRequestBaseForm):
 
     class Meta:
         model = ContactRequest
-        fields = ('first_name', 'last_name', 'company_name', 'email', 'message', 'contact_type', 'date', 'time')
-
-    def __init__(self, *args, **kwargs):
-        super(ContactRequestForm, self).__init__(*args, **kwargs)
-
-        #  required fiels
-        for key in self.fields:
-            if key != 'message' and key != 'type' and key != 'date' and key != 'time':
-                self.fields[key].required = True
+        fields = ('salutation', 'first_name', 'last_name', 'company_name', 'contact_type', 'email', 'message', 'phone', 'date', 'time')