+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The version numbers referring to the version in [allink_core](git@github.com/all
 - field "active" renamed to "is_active" on all models
 - category nav in case no categories are selected fixed
 - Clases inherting from Base Model: Separated filter and categories fieldsets, applied default ordering on querysets
+- Made opening hours field in locations app translatable
 
 ## v0.0.8